requirements.txt
asgiref==3.2.5
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.4
django-ckeditor==5.9.0
django-import-export==2.0.2
django-js-asset==1.2.2
django-tinymce==2.8.0
django-tinymce4==3.0.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==8.1.0
psycopg2==2.8.6
pytz==2019.3
PyYAML==5.3.1
sqlparse==0.3.1
tablib==1.1.0
xlrd==1.2.0
xlwt==1.3.0
